indicator: Add solid style indicator
This alternative style of the indicator removes the legacy glow effect
around the dot (the indicator) and makes it easy to customize the dot's
color.

In addition, it is also trying to come up with better positioning of the
dots to suit the style. So, the indicators will be positioned evenly
between the bottom offset of the dock (`theme.get_bottom_offset ()`) and
the bottom offset of the icon container.