WIP
Revert commit 581819f5d539210e1ec282d72a0c0c87aec926ed

The urgent bounce height (when zoom-enabled and not hovering on the
dock/icon) stays the same as when the icon is zoomed.
Thus, need to update cache of the bounce height when not hovering
on the hover area of the dock. But currently don't know where to do so.

Or maybe we can do the opposite way. Currently we set the bounce height
in the initial process regarding to the value of zoomed icon size
instead of just the icon size. I think instead we need to let the
initial process set the bounce height using only the normal icon size.
And then in hover state we can adjust the bounce height regarding to
the zoomed icon size.